# docs/swagger.yaml
basePath: /
definitions:
  dtos.AccumulatedRewardRequest:
    properties:
      cashback_accumulated:
        type: number
      points_accumulated:
        type: number
      store_id:
        type: integer
      user_id:
        type: integer
    type: object
  dtos.BranchRequest:
    properties:
      address:
        type: string
      name:
        type: string
      store_id:
        type: integer
    type: object
  dtos.CampaignRequest:
    properties:
      branch_id:
        type: integer
      end_date:
        type: string
      name:
        type: string
      percentage:
        type: number
      start_date:
        type: string
      type:
        type: string
    type: object
...
info:
  contact: {}
  description: API test.
  title: mi api
  version: "1.0"
paths:
  /leal-test/acumulaterewards:
    get:
      consumes:
      - application/json
      description: Get all accumulated rewards
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get all accumulated rewards
      tags:
      - accumulated_rewards
    post:
      consumes:
      - application/json
      description: Create a new accumulated reward
      parameters:
      - description: Reward data
        in: body
        name: reward
        required: true
        schema:
          $ref: '#/definitions/dtos.AccumulatedRewardRequest'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Create a new accumulated reward
      tags:
      - accumulated_rewards
  /leal-test/acumulaterewards/{id}:
    delete:
      description: Delete an accumulated reward by ID
      parameters:
      - description: Reward ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Delete accumulated reward by ID
      tags:
      - accumulated_rewards
    get:
      consumes:
      - application/json
      description: Get accumulated reward by ID
      parameters:
      - description: Reward ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get accumulated reward by ID
      tags:
      - accumulated_rewards
    put:
      consumes:
      - application/json
      description: Update an accumulated reward
      parameters:
      - description: Reward ID
        in: path
        name: id
        required: true
        type: integer
      - description: Reward data
        in: body
        name: reward
        required: true
        schema:
          $ref: '#/definitions/dtos.AccumulatedRewardRequest'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Update an accumulated reward
      tags:
      - accumulated_rewards
  /leal-test/acumulaterewards/user/{user_id}/store/{store_id}:
    get:
      consumes:
      - application/json
      description: Get accumulated reward by UserID and StoreID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Store ID
        in: path
        name: store_id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get accumulated reward by UserID and StoreID
      tags:
      - accumulated_rewards
  /leal-test/branches:
    get:
      consumes:
      - application/json
      description: Get all branches
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get all branches
      tags:
      - branches
    post:
      consumes:
      - application/json
      description: Create a new branch
      parameters:
      - description: Branch data
        in: body
        name: branch
        required: true
        schema:
          $ref: '#/definitions/dtos.BranchRequest'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Create a new branch
      tags:
      - branches
  /leal-test/branches/{id}:
    delete:
      description: Delete a branch by ID
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Delete branch by ID
      tags:
      - branches
    get:
      consumes:
      - application/json
      description: Get branch by ID
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get branch by ID
      tags:
      - branches
    put:
      consumes:
      - application/json
      description: Update a branch
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      - description: Branch data
        in: body
        name: branch
        required: true
        schema:
          $ref: '#/definitions/dtos.BranchRequest'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Update a branch
      tags:
      - branches
  /leal-test/campaigns:
    get:
      consumes:
      - application/json
      description: Get all campaigns
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get all campaigns
      tags:
      - campaigns
    post:
      consumes:
      - application/json
      description: Create a new campaign
      parameters:
      - description: Campaign data
        in: body
        name: campaign
        required: true
        schema:
          $ref: '#/definitions/dtos.CampaignRequest'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Create a new campaign
      tags:
      - campaigns
  /leal-test/campaigns/{id}:
    delete:
      description: Delete a campaign by ID
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Delete campaign by ID
      tags:
      - campaigns
    get:
      consumes:
      - application/json
      description: Get campaign by ID
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get campaign by ID
      tags:
      - campaigns
    put:
      consumes:
      - application/json
      description: Update a campaign
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: integer
      - description: Campaign data
        in: body
        name: campaign
        required: true
        schema:
          $ref: '#/definitions/dtos.CampaignRequest'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Update a campaign
      tags:
      - campaigns
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"